feat(创建实体类):完成学习计划表中第一步：实体表(商品管理表和订单管理表)设计与实现
