+++ practice-project/pom.xml
@@ -8,7 +8,7 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.jetlinks</groupId>
+    <groupId>org.jetlinks.community</groupId>
     <artifactId>practice-project</artifactId>
     <version>2.1.0-SNAPSHOT</version>
     <name>practice-project</name>
