+++ practice-project/src/main/java/org/jetlinks/community/example/entity/ProductEntity.java
@@ -1,56 +1,54 @@
 package org.jetlinks.community.example.entity;
 
+import org.hswebframework.ezorm.rdb.mapping.annotation.ColumnType;
+import org.hswebframework.ezorm.rdb.mapping.annotation.Comment;
+import org.hswebframework.ezorm.rdb.mapping.annotation.EnumCodec;
+import org.hswebframework.web.api.crud.entity.GenericEntity;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.Setter;
-import org.hswebframework.ezorm.rdb.mapping.annotation.ColumnType;
 import org.hswebframework.ezorm.rdb.mapping.annotation.DefaultValue;
-import org.hswebframework.ezorm.rdb.mapping.annotation.EnumCodec;
-import org.hswebframework.web.api.crud.entity.GenericEntity;
 import org.hswebframework.web.api.crud.entity.RecordCreationEntity;
 import org.hswebframework.web.api.crud.entity.RecordModifierEntity;
 import org.hswebframework.web.crud.annotation.EnableEntityEvent;
 import org.hswebframework.web.crud.generator.Generators;
-import org.hswebframework.web.utils.DigestUtils;
 import org.hswebframework.web.validator.CreateGroup;
 import org.jetlinks.community.example.enums.OrderType;
+import org.jetlinks.community.example.enums.ProductType;
 
 
 import javax.persistence.Column;
 import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Pattern;
 import java.sql.JDBCType;
 
 /**
- * 订单管理实体类
+ * 商品管理实体类
  *
  * @author wangyujie
  * @since 1.0
  */
-@Table(name = "s_order")
+
+@Table(name = "s_product")
 @Getter
 @Setter
-@Schema(description = "订单管理表")
+@Comment("商品实体类")
 @EnableEntityEvent //开启实体类crud事件
-public class Order extends GenericEntity<String>
+public class ProductEntity extends GenericEntity<String>
 //实现这2个接口标记此实体类需要记录创建人修改人信息，在crud时会自动填充对应的信息。
     implements RecordCreationEntity, RecordModifierEntity {
 
-    //订单流水号自动生成，不可修改
-    @Column(updatable = false)
-    @Schema(description = "订单流水号")
-    @GeneratedValue
-    private  String orderSerialNumber;
-
-
-
     //平台ID长度统一64,商品id,不可修改
+    @Id
     @Column(length = 64, nullable = false, updatable = false)
     @NotBlank(groups = CreateGroup.class)
     @Schema(description = "商品ID", accessMode = Schema.AccessMode.READ_ONLY)
-    private String productId;
-
+    public String getId() {
+        return super.getId();
+    }
 
     @Column(length = 64, nullable = false)
     @NotBlank(groups = CreateGroup.class)
@@ -58,16 +56,27 @@ public class Order extends GenericEntity<String>
     private String name;
 
     @Column
+    @Pattern(regexp = "[a-zA-Z]{3}@[0-9]{2}", message = "3位字母编码@2为数字编码，例如abc@12")
     @Schema(description = "商品批次")
     private String productBatch;
+
 
 
     @Column
     //使用枚举掩码来存储多选值,因此数据库中用bigint来存储
     @EnumCodec
     @ColumnType(javaType = Long.class, jdbcType = JDBCType.BIGINT)
-    @Schema(description = "订单类型")
-    private OrderType orderType;
+    @Schema(description = "商品分类")
+    private ProductType productType;
+
+
+    @GeneratedValue
+    @Schema(description = "上架日期", accessMode = Schema.AccessMode.READ_ONLY)
+    private Long upTime;
+
+
+    @Schema(description = "下架日期", accessMode = Schema.AccessMode.READ_ONLY)
+    private Long downTime;
 
     //平台ID长度统一64,创建人不为空,不可修改
     @Column(length = 64, nullable = false, updatable = false)
@@ -87,22 +96,5 @@ public class Order extends GenericEntity<String>
     @DefaultValue(generator = Generators.CURRENT_TIME)
     @Schema(description = "创建时间", accessMode = Schema.AccessMode.READ_ONLY)
     private Long createTime;
-
-    @Override
-    public String getId() {
-        if (super.getId() == null) {
-            generateId();
-        }
-        return super.getId();
-    }
-
-    public void generateId() {
-        String id = generateOrderId(orderSerialNumber, productId);
-        setId(id);
-    }
-
-    public static String generateOrderId(String orderSerialNumber, String productId) {
-        return DigestUtils.md5Hex(String.join(orderSerialNumber, "|", productId));
-    }
 
 }
